fixup test harness
