+++ src/Microsoft.VisualStudio.ProjectSystem.Managed.VS.UnitTests/ProjectSystem/VS/Rename/RenamerTestsBase.cs
@@ -1,5 +1,8 @@
 ﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
+using System;
+using System.Linq;
+using System.Threading.Tasks;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.Text;
 
@@ -35,6 +38,25 @@ namespace Microsoft.VisualStudio.ProjectSystem.VS.Rename
                                 filePath: fileName)
                          })
                 });
+        }
+
+        internal async Task RenameAsync(string sourceCode, string oldFilePath, string newFilePath, IUserNotificationServices userNotificationServices, IRoslynServices roslynServices, string language)
+        {
+            using var ws = new AdhocWorkspace();
+            var projectId = ProjectId.CreateNewId();
+            Solution solution = ws.AddSolution(InitializeWorkspace(projectId, newFilePath, sourceCode, language));
+            Project project = (from d in solution.Projects where d.Id == projectId select d).FirstOrDefault();
+
+            var unconfiguredProject = UnconfiguredProjectFactory.Create(filePath: $@"C:\project1.{ProjectFileExtension}");
+            var projectServices = IUnconfiguredProjectVsServicesFactory.Implement(
+                    threadingServiceCreator: () => IProjectThreadingServiceFactory.Create(),
+                    unconfiguredProjectCreator: () => unconfiguredProject);
+            var unconfiguredProjectTasksService = IUnconfiguredProjectTasksServiceFactory.Create();
+            var environmentOptionsFactory = IEnvironmentOptionsFactory.Implement((string category, string page, string property, bool defaultValue) => { return true; });
+
+            var renamer = new Renamer(projectServices, unconfiguredProjectTasksService, ws, environmentOptionsFactory, userNotificationServices, roslynServices);
+            await renamer.HandleRenameAsync(oldFilePath, newFilePath)
+                         .TimeoutAfter(TimeSpan.FromSeconds(1));
         }
     }
 }
